requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
boto3==1.38.36
    # via -r requirements.in
botocore==1.38.36
    # via
    #   boto3
    #   s3transfer
certifi==2025.6.15
    # via influxdb-client
colorama==0.4.6
    # via loguru
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.58.4
    # via matplotlib
influxdb-client==1.49.0
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kiwisolver==1.4.8
    # via matplotlib
loguru==0.7.3
    # via -r requirements.in
matplotlib==3.10.3
    # via -r requirements.in
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
    #   scipy
packaging==25.0
    # via matplotlib
pandas==2.2.3
    # via -r requirements.in
pika==1.3.2
    # via -r requirements.in
pillow==11.2.1
    # via matplotlib
prometheus-client==0.22.1
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   influxdb-client
    #   matplotlib
    #   pandas
python-dotenv==1.1.0
    # via -r requirements.in
pytz==2025.2
    # via pandas
reactivex==4.0.4
    # via influxdb-client
s3transfer==0.13.0
    # via boto3
scipy==1.15.3
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
typing-extensions==4.14.0
    # via reactivex
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via
    #   botocore
    #   influxdb-client
win32-setctime==1.2.0
    # via loguru

# The following packages are considered to be unsafe in a requirements file:
# setuptools
